benford_backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cycler==0.10.0
Django==3.1.1
djangorestframework==3.11.1
install==1.3.3
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
